+++ CHANGELOG.md
@@ -16,8 +16,6 @@ This project should adhere to [Semantic Versioning](https://semver.org/spec/v2.0
 
 ### Added
 
-* Documentation has been through a major overhaul.
-* Added some information on how to connect to Google in the doc and examples.
 * `event.component` is now an alias for `event.icalendar_component`.
 * `get_davclient` (earlier called `auto_conn`) is more complete now - https://github.com/python-caldav/caldav/pull/502 - https://github.com/python-caldav/caldav/issues/485 - https://github.com/python-caldav/caldav/pull/507
   * It can read from environment (including environment variable for reading from test config and for locating the config file).
@@ -26,6 +24,12 @@ This project should adhere to [Semantic Versioning](https://semver.org/spec/v2.0
   * Improved tests (but no test for inheritance yet).
   * Documentation, linked up from the reference section of the doc.
   * It's allowable with a yaml config file, but the yaml module is not included in the dependencies yet ... so late imports as for now, and the import is wrapped in a try/except-block
+* New method `davclient.principals()` will return all principals on server - if server permits.
+
+### Documentation and examples
+
+* Documentation has been through a major overhaul.
+* Added some information on how to connect to Google in the doc and examples.
 * Looked through and brushed up the examples, two of them are now executed by the unit tests.  Added a doc section on the examples.
 
 ### Fixes